refactor(tasks): move X2 tracking owner [skip ci]
