+++ src/unilab/tasks/motion_tracking/__init__.py
@@ -1,0 +1,15 @@
+"""Motion-tracking task packages."""
+
+from .x2 import (
+    X2MotionTrackingCfg,
+    X2WallFlipTrackingCfg,
+    X2WallFlipTrackingEnv,
+    X2WallFlipTrackingEnvCfg,
+)
+
+__all__ = [
+    "X2MotionTrackingCfg",
+    "X2WallFlipTrackingCfg",
+    "X2WallFlipTrackingEnv",
+    "X2WallFlipTrackingEnvCfg",
+]
